refactor: clean up tobii-service codebase
- Apply Copilot review suggestions on PR #11:
  - Fix docstring to mention all 3 frozen-exe fixes
  - Use os.devnull instead of hardcoded platform check
  - Wrap create_app() in try-except for better diagnostics
  - Document intentional devnull file handle lifetime

- Modernize Pydantic APIs:
  - Replace deprecated class Config with model_config/ConfigDict
  - Replace .dict() with .model_dump()
  - Replace typing.Optional/List/Dict/Tuple with built-in equivalents

- Split monolithic widgets.py (316 lines) into gui/widgets/ package:
  - header.py, status_card.py, control_buttons.py, exit_button.py, confirm_dialog.py

- Remove unused Styles.configure_styles() method (ttk, not used with CustomTkinter)
- Remove redundant docstrings and comments
- Remove outdated file structure section from API.md
- Fix GitHub username in docs (Khalil-Elemam -> khalilelemam)
- Run ruff format and lint (I,UP,F rules)
