+++ tobii-service/gui/widgets/__init__.py
@@ -1,0 +1,15 @@
+"""GUI widget components for Lexora Eye Tracker Service."""
+
+from gui.widgets.confirm_dialog import ConfirmDialog
+from gui.widgets.control_buttons import ControlButtons
+from gui.widgets.exit_button import ExitButton
+from gui.widgets.header import HeaderWidget
+from gui.widgets.status_card import StatusCard
+
+__all__ = [
+    "HeaderWidget",
+    "StatusCard",
+    "ControlButtons",
+    "ExitButton",
+    "ConfirmDialog",
+]
